Move OpenFileJob to its own header
